Update 2 - electric boogaloo
Updated the main script so that it was more consistent as it would sometimes get stuck in a loop. Now the script send emails regardless of if there is a change of weather and instead replaces the main body of the email if there is no data. This may be changed in future.